requirements.txt
opencv==4.6.0
pillow==10.0.1